# sm3_mb optimization by SVE2 for aarch64