Test result on N2/V1 micro-architecture:

+--------+-------------+-------------+------------+
|        | CPU1 (N2)   | CPU2 (N2)   | CPU3 (V1)  |
|        | 128-bit VL  | 128-bit VL  | 256-bit VL |
+--------+-------------+-------------+------------+
|  neon  |  397 MB/s   |  495 MB/s   |  464 MB/s  |
+--------+-------------+-------------+------------+
|  sve2  |  425 MB/s   |  516 MB/s   |            |
|        |   (+7%)     |  (+4.2%)    |
+--------+-------------+-------------+------------+
|  sve   |             |             |  652 MB/s  |
|        |             |             |    (+40%)  |
+--------+-------------+-------------+------------+

Change-Id: Ie73098d45a83eb438cdf3121d97d10e477ea9665
Signed-off-by: Daniel Hu <Daniel.Hu@arm.com>